+++ sotodlib/mapmaking/__init__.py
@@ -1,7 +1,3 @@
-from .bench import *
-from .log import *
-from .mapmakers import demod_mapmaker, ml_mapmaker
-from .noise_model import *
-from .obs_grouping import *
-from .pointing_matrix import *
-from .utils import utils
+from .ml_mapmaker import SignalCut, SignalMap, SignalSrcsamp, MLMapmaker # noqa: F401 
+from .noise_model import read_nmat, write_nmat, NmatUncorr, NmatDetvecs, Nmat, NmatWhite # noqa: F401
+from .demod_mapmaker import DemodMapmaker, DemodSignalMap # noqa: F401
