--- sotodlib/mapmaking/mapmakers/__init__.py
@@ -1,1 +1,0 @@
-from .demod_mapmaker import DemodMapmaker, make_demod_map  # noqa: F401
