+++ sotodlib/mapmaking/obs_grouping.py
@@ -22,7 +22,6 @@ from scipy import ndimage
 
 from .utils import get_ids
 
-
 class NoTODFound(Exception):
     def __init__(self, msg):
         self.msg = msg
