+++ sotodlib/mapmaking/utils/__init__.py
@@ -1,3 +1,24 @@
-from .utils import (MultiZipper, atomic_db_aux,  # noqa: F401
-                    evaluate_recentering, import_optional,
-                    recentering_to_quat_lonlat, unarr)
+from .utils import (
+    import_optional,  # noqa: F401
+    get_ids,  # noqa: F401
+    parse_recentering,  # noqa: F401
+    get_subids,  # noqa: F401
+    filter_subids,  # noqa: F401
+    setup_passes,  # noqa: F401
+    fix_boresight_glitches,  # noqa: F401
+    find_usable_detectors,  # noqa: F401
+    inject_map,  # noqa: F401
+    downsample_obs,  # noqa: F401
+    resample,  # noqa: F401
+    MultiZipper,  # noqa: F401
+    ArrayZipper,  # noqa: F401
+    MapZipper,  # noqa: F401
+    unarr,  # noqa: F401
+    safe_invert_div,  # noqa: F401
+    apply_window,  # noqa: F401
+    makebins,  # noqa: F401
+    measure_detvecs,  # noqa: F401
+    find_modes_jon,  # noqa: F401
+    woodbury_invert,  # noqa: F401
+)  # noqa: F401
+from .log import DEBUG, init  # noqa: F401
